code cleanup and added modifier areas
