--- scenes/2,22-virtual-cameras/src/components.ts
@@ -1,10 +1,0 @@
-import { Schemas, engine } from '@dcl/sdk/ecs'
-
-
-// We use this component to track and group all spinning entities.
-// engine.getEntitiesWith(Spinner)
-export const Spinner = engine.defineComponent('spinner', { speed: Schemas.Number })
-
-// We use this component to track and group all the cubes.
-// engine.getEntitiesWith(Cube)
-export const Cube = engine.defineComponent('cube-id', {})
