--- scenes/2,22-virtual-cameras/src/factory.ts
@@ -1,42 +1,0 @@
-import {
-  Entity,
-  engine,
-  Transform,
-  MeshRenderer,
-  MeshCollider,
-  PointerEvents,
-  PointerEventType,
-  InputAction,
-  Material
-} from '@dcl/sdk/ecs'
-import { Cube, Spinner } from './components'
-import { Color4 } from '@dcl/sdk/math'
-import { getRandomHexColor } from './utils'
-
-// Cube factory
-export function createCube(x: number, y: number, z: number, spawner = true): Entity {
-  const entity = engine.addEntity()
-
-  // Used to track the cubes
-  Cube.create(entity)
-
-  Transform.create(entity, { position: { x, y, z } })
-
-  // set how the cube looks and collides
-  MeshRenderer.setBox(entity)
-  MeshCollider.setBox(entity)
-  Material.setPbrMaterial(entity, { albedoColor: Color4.fromHexString(getRandomHexColor()) })
-
-  // Make the cube spin, with the circularSystem
-  Spinner.create(entity, { speed: 100 * Math.random() })
-
-  // Create PointerEvent with the hover feedback.
-  // We are going to check the onClick event on the changeColorSystem.
-  PointerEvents.create(entity, {
-    pointerEvents: [
-      { eventType: PointerEventType.PET_DOWN, eventInfo: { button: InputAction.IA_POINTER, hoverText: 'Change Color' } }
-    ]
-  })
-
-  return entity
-}
