--- scenes/2,22-virtual-cameras/src/ui.tsx
@@ -1,79 +1,0 @@
-import {
-  engine,
-  Transform,
-} from '@dcl/sdk/ecs'
-import { Color4 } from '@dcl/sdk/math'
-import ReactEcs, { Button, Label, ReactEcsRenderer, UiEntity } from '@dcl/sdk/react-ecs'
-import { Cube } from './components'
-import { createCube } from './factory'
-
-export function setupUi() {
-  ReactEcsRenderer.setUiRenderer(uiComponent)
-}
-
-const uiComponent = () => (
-  <UiEntity
-    uiTransform={{
-      width: 400,
-      height: 230,
-      margin: '16px 0 8px 270px',
-      padding: 4,
-    }}
-    uiBackground={{ color: Color4.create(0.5, 0.8, 0.1, 0.6) }}
-  >
-    <UiEntity
-      uiTransform={{
-        width: '100%',
-        height: '100%',
-        flexDirection: 'column',
-        alignItems: 'center',
-        justifyContent: 'space-between'
-      }}
-      uiBackground={{ color: Color4.fromHexString("#70ac76ff") }}
-    >
-      <UiEntity
-        uiTransform={{
-          width: '100%',
-          height: 50,
-          margin: '8px 0'
-        }}
-        uiBackground={{
-          textureMode: 'center',
-          texture: {
-            src: 'images/scene-thumbnail.png',
-          },
-        }}
-        uiText={{ value: 'SDK7', fontSize: 18 }}
-      />
-      <Label
-        onMouseDown={() => {console.log('Player Position clicked !')}}
-        value={`Player: ${getPlayerPosition()}`}
-        fontSize={18}
-        uiTransform={{ width: '100%', height: 30 } }
-      />
-      <Label
-        onMouseDown={() => {console.log('# Cubes clicked !')}}
-        value={`# Cubes: ${[...engine.getEntitiesWith(Cube)].length}`}
-        fontSize={18}
-        uiTransform={{ width: '100%', height: 30 } }
-      />
-      <Button
-        uiTransform={{ width: 100, height: 40, margin: 8 }}
-        value='Spawn cube'
-        variant='primary'
-        fontSize={14}
-        onMouseDown={() => {
-          createCube(1 + Math.random() * 8, Math.random() * 8, 1 + Math.random() * 8, false)
-        }}
-      />
-     </UiEntity>
-  </UiEntity>
-)
-
-function getPlayerPosition() {
-  const playerPosition = Transform.getOrNull(engine.PlayerEntity)
-  if (!playerPosition) return ' no data yet'
-  const { x, y, z } = playerPosition.position
-  return `{X: ${x.toFixed(2)}, Y: ${y.toFixed(2)}, z: ${z.toFixed(2)} }`
-}
-
