--- scenes/2,22-virtual-cameras/src/utils.ts
@@ -1,9 +1,0 @@
-export function getRandomHexColor(): string {
-  const letters = "0123456789ABCDEF";
-  let color = "#";
-  for (let i = 0; i < 6; i++) {
-    color += letters[Math.floor(Math.random() * 16)];
-  }
-  return color;
-}
-
